add mir::type_check, extract mir::lower;
