+++ ergoscript-compiler/src/mir/type_check.rs
@@ -1,0 +1,17 @@
+use ergo_lib::ast::expr::Expr;
+
+#[derive(Debug, PartialEq)]
+pub struct TypeCheckError {}
+
+pub fn type_check(e: Expr) -> Result<Expr, TypeCheckError> {
+    match &e {
+        Expr::BinOp(bin) => {
+            if bin.left.tpe() == bin.right.tpe() {
+                Ok(e)
+            } else {
+                Err(TypeCheckError {})
+            }
+        }
+        _ => Ok(e),
+    }
+}
